Refactor: have the lifted hooks read their types from the api layer
The two hooks that moved up were still importing their types through a page's
re-export shim, which left the arrow pointing uphill for types the api module
already exports. The pages keep those files for their form types.
